requirements.txt
python-dotenv
requests
yt-dlp
